Add support for an optional `timeout` parameter
This changes the signature of `acquireAsync` to `acquireAsync({ timeout?: number })`. When specified, the returned promise is rejected if the lock cannot be acquired before the timeout elapses.

Added unit tests to test subtle behavior around timing and the event loop. Namely, if the lock is freed earlier in the same iteration of the event loop as the timer firing, the lock should be acquired. Likewise, if the timer fires right before the lock is released during the same turn of the event loop, the promise is rejected. Unit tests have 100% branch statement coverage over these tricky cases.