Change the waiting acquirer queue from an array to a Set
We need to delete entries from the acquirer queue. A Set allows for faster random access than a linear scan of the array, plus potential linear splicing of the array.

This works because Sets preserve insertion order so they can be used like a queue. Added a unit test that checks that the order of the acquireAsync calls is honored for fair scheduling.